update getTokenInfo to be plural
