+++ packages/neon-api/__tests__/api/getTokenInfos.ts
@@ -1,7 +1,7 @@
 import { rpc } from "@cityofzion/neon-core";
-import { getTokenInfo } from "../../src/api/getTokenInfo";
+import { getTokenInfos } from "../../src/api/getTokenInfos";
 
-describe("getTokenInfo", () => {
+describe("getTokenInfos", () => {
   test("success", async () => {
     const mockClient = ({
       invokeScript: jest.fn(async () => {
@@ -27,22 +27,49 @@ describe("getTokenInfo", () => {
               type: "Integer",
               value: "3000006384920100",
             },
+            {
+              type: "ByteString",
+              value: "TkVP",
+            },
+            {
+              type: "ByteString",
+              value: "bmVv",
+            },
+            {
+              type: "Integer",
+              value: "0",
+            },
+            {
+              type: "Integer",
+              value: "100000000",
+            },
           ],
         };
       }),
     } as unknown) as rpc.NeoServerRpcClient;
 
-    const result = await getTokenInfo(
-      "668e0c1f9d7b70a99dd9e06eadd4c784d641afbc",
+    const result = await getTokenInfos(
+      [
+        "668e0c1f9d7b70a99dd9e06eadd4c784d641afbc",
+        "de5f57d430d3dece511cf975a8d37848cb9e0525",
+      ],
       mockClient
     );
 
-    expect(result).toStrictEqual({
-      name: "GAS",
-      symbol: "gas",
-      decimals: 8,
-      totalSupply: "30000063.84920100",
-    });
+    expect(result).toStrictEqual([
+      {
+        name: "GAS",
+        symbol: "gas",
+        decimals: 8,
+        totalSupply: "30000063.84920100",
+      },
+      {
+        name: "NEO",
+        symbol: "neo",
+        decimals: 0,
+        totalSupply: "100000000",
+      },
+    ]);
   });
 
   test("VM fault", async () => {
@@ -60,8 +87,8 @@ describe("getTokenInfo", () => {
 
     expect(
       async () =>
-        await getTokenInfo(
-          "668e0c1f9d7b70a99dd9e06eadd4c784d641afbc",
+        await getTokenInfos(
+          ["668e0c1f9d7b70a99dd9e06eadd4c784d641afbc"],
           mockClient
         )
     ).rejects.toThrow("expected exception message");
@@ -81,8 +108,8 @@ describe("getTokenInfo", () => {
 
     expect(
       async () =>
-        await getTokenInfo(
-          "668e0c1f9d7b70a99dd9e06eadd4c784d641afbc",
+        await getTokenInfos(
+          ["668e0c1f9d7b70a99dd9e06eadd4c784d641afbc"],
           mockClient
         )
     ).rejects.toThrow("No exception returned");
