+++ packages/neon-api/src/api/index.ts
@@ -1,2 +1,2 @@
 export * from "./getTokenBalances";
-export * from "./getTokenInfo";
+export * from "./getTokenInfos";
